♻️  zb: move NonceTCP connection test
